+++ crates/gosub_lattice/src/sizing/columns.rs
@@ -48,9 +48,9 @@ pub fn column_specs<T: TableTree>(tree: &T, model: &TableModel<T::NodeId>) -> Ve
 pub fn compute_column_widths<T: TableTree>(
     tree: &mut T,
     n_cols: usize,
-    explicit_table_width: Option<f32>,
-    available_width: f32,
-    border_spacing_x: f32,
+    explicit_table_width: Option<f64>,
+    available_width: f64,
+    border_spacing_x: f64,
     grids: &[&SectionGrid<T::NodeId>],
     sizing: TableSizing,
     col_specs: &[CssLength],
@@ -58,14 +58,14 @@ pub fn compute_column_widths<T: TableTree>(
     // CAPMIN: the caption's minimum border-box width. CSS 2 §17.5.2 makes the used
     // table width the greater of the width the columns require and CAPMIN, in both
     // layout modes. 0.0 when there is no caption.
-    capmin: f32,
-) -> (Vec<f32>, f32) {
+    capmin: f64,
+) -> (Vec<f64>, f64) {
     if n_cols == 0 {
         return (Vec::new(), capmin.max(0.0));
     }
 
     // Total space consumed by border-spacing gutters.
-    let spacing_total = (n_cols as f32 + 1.0) * border_spacing_x;
+    let spacing_total = (n_cols as f64 + 1.0) * border_spacing_x;
 
     if sizing == TableSizing::Fixed {
         let table_width = explicit_table_width.unwrap_or(available_width).max(capmin);
@@ -78,9 +78,9 @@ pub fn compute_column_widths<T: TableTree>(
     // otherwise against the containing block.
     let percent_basis = explicit_table_width.unwrap_or(available_width);
 
-    let mut min = vec![0.0_f32; n_cols];
-    let mut max = vec![0.0_f32; n_cols];
-    let mut spec: Vec<Option<f32>> = vec![None; n_cols];
+    let mut min = vec![0.0_f64; n_cols];
+    let mut max = vec![0.0_f64; n_cols];
+    let mut spec: Vec<Option<f64>> = vec![None; n_cols];
 
     for (i, s) in col_specs.iter().take(n_cols).enumerate() {
         if let Some(px) = s.resolve(percent_basis) {
@@ -90,7 +90,7 @@ pub fn compute_column_widths<T: TableTree>(
 
     // Single-column cells contribute directly; colspan cells are collected and
     // distributed afterwards, shortest spans first.
-    let mut spanners: Vec<(usize, usize, f32, f32, Option<f32>)> = Vec::new();
+    let mut spanners: Vec<(usize, usize, f64, f64, Option<f64>)> = Vec::new();
     for grid in grids {
         for cell in grid.cells() {
             let (min_c, max_c) = tree.cell_intrinsic_widths(cell.node);
@@ -116,7 +116,7 @@ pub fn compute_column_widths<T: TableTree>(
             continue;
         }
         // The spanning cell runs across the gutters between its columns.
-        let gutters = border_spacing_x * range.len().saturating_sub(1) as f32;
+        let gutters = border_spacing_x * range.len().saturating_sub(1) as f64;
         distribute_deficit(&mut min, range.clone(), min_c - gutters, &max);
         distribute_deficit(&mut max, range.clone(), max_c - gutters, &min);
         // A specified width on a spanning cell divides evenly over columns
@@ -124,7 +124,7 @@ pub fn compute_column_widths<T: TableTree>(
         if let Some(w) = w {
             let open: Vec<usize> = range.clone().filter(|&c| spec[c].is_none()).collect();
             if open.len() == range.len() {
-                let share = (w - gutters) / open.len() as f32;
+                let share = (w - gutters) / open.len() as f64;
                 for c in open {
                     spec[c] = Some(share.max(0.0));
                 }
@@ -134,12 +134,12 @@ pub fn compute_column_widths<T: TableTree>(
 
     // A specified width cannot shrink a column below its min-content width; a
     // specified column contributes that fixed width as both its min and max.
-    let contrib_min: Vec<f32> = (0..n_cols).map(|c| spec[c].map_or(min[c], |s| s.max(min[c]))).collect();
-    let contrib_max: Vec<f32> = (0..n_cols)
+    let contrib_min: Vec<f64> = (0..n_cols).map(|c| spec[c].map_or(min[c], |s| s.max(min[c]))).collect();
+    let contrib_max: Vec<f64> = (0..n_cols)
         .map(|c| spec[c].map_or(max[c].max(min[c]), |s| s.max(min[c])))
         .collect();
-    let cmin: f32 = contrib_min.iter().sum();
-    let cmax: f32 = contrib_max.iter().sum();
+    let cmin: f64 = contrib_min.iter().sum();
+    let cmax: f64 = contrib_max.iter().sum();
 
     // With no intrinsic information at all shrink-to-fit would collapse the table -
     // fill the available width instead. Only for mock trees using the intrinsics
@@ -163,8 +163,8 @@ pub fn compute_column_widths<T: TableTree>(
     let mut widths = contrib_min.clone();
     let mut extra = inner - cmin;
     if extra > 0.0 {
-        let growth: Vec<f32> = (0..n_cols).map(|c| contrib_max[c] - contrib_min[c]).collect();
-        let growth_total: f32 = growth.iter().sum();
+        let growth: Vec<f64> = (0..n_cols).map(|c| contrib_max[c] - contrib_min[c]).collect();
+        let growth_total: f64 = growth.iter().sum();
         if growth_total > 0.0 {
             let g = extra.min(growth_total);
             for c in 0..n_cols {
@@ -177,13 +177,13 @@ pub fn compute_column_widths<T: TableTree>(
             // their max-content width so content-heavy columns absorb more.
             let autos: Vec<usize> = (0..n_cols).filter(|&c| spec[c].is_none()).collect();
             let targets = if autos.is_empty() { (0..n_cols).collect() } else { autos };
-            let weight_total: f32 = targets.iter().map(|&c| contrib_max[c]).sum();
+            let weight_total: f64 = targets.iter().map(|&c| contrib_max[c]).sum();
             if weight_total > 0.0 {
                 for &c in &targets {
                     widths[c] += extra * contrib_max[c] / weight_total;
                 }
             } else {
-                let equal = extra / targets.len() as f32;
+                let equal = extra / targets.len() as f64;
                 for &c in &targets {
                     widths[c] += equal;
                 }
@@ -197,19 +197,19 @@ pub fn compute_column_widths<T: TableTree>(
 /// Raise the values in `range` so they sum to at least `required`. The deficit
 /// is split proportionally to `weights` (content-heavy columns absorb more),
 /// equally when the weights are all zero.
-fn distribute_deficit(vals: &mut [f32], range: std::ops::Range<usize>, required: f32, weights: &[f32]) {
-    let current: f32 = vals[range.clone()].iter().sum();
+fn distribute_deficit(vals: &mut [f64], range: std::ops::Range<usize>, required: f64, weights: &[f64]) {
+    let current: f64 = vals[range.clone()].iter().sum();
     if required <= current {
         return;
     }
     let deficit = required - current;
-    let weight_total: f32 = weights[range.clone()].iter().sum();
+    let weight_total: f64 = weights[range.clone()].iter().sum();
     if weight_total > 0.0 {
         for c in range {
             vals[c] += deficit * weights[c] / weight_total;
         }
     } else {
-        let equal = deficit / range.len() as f32;
+        let equal = deficit / range.len() as f64;
         for c in range {
             vals[c] += equal;
         }
@@ -223,12 +223,12 @@ fn distribute_deficit(vals: &mut [f32], range: std::ops::Range<usize>, required:
 fn fixed_column_widths<T: TableTree>(
     tree: &T,
     n_cols: usize,
-    available: f32,
+    available: f64,
     grids: &[&SectionGrid<T::NodeId>],
     col_specs: &[CssLength],
     collapsed_borders: &HashMap<T::NodeId, CollapsedBorders>,
-) -> Vec<f32> {
-    let mut explicit: Vec<Option<f32>> = vec![None; n_cols];
+) -> Vec<f64> {
+    let mut explicit: Vec<Option<f64>> = vec![None; n_cols];
 
     // Percentages resolve against the space actually available to columns:
     // the table width minus the border-spacing gutters (CSS 2 §17.5.2.1
@@ -255,7 +255,7 @@ fn fixed_column_widths<T: TableTree>(
                 let border = effective_border(tree, cell.node, collapsed_borders);
                 let padding = read_padding(tree, cell.node);
                 let outer = w + padding.left + padding.right + border.left + border.right;
-                let share = outer / cell.colspan as f32;
+                let share = outer / cell.colspan as f64;
                 let end = (cell.col + cell.colspan).min(n_cols);
                 for slot in &mut explicit[cell.col..end] {
                     slot.get_or_insert(share);
@@ -267,13 +267,13 @@ fn fixed_column_widths<T: TableTree>(
         }
     }
 
-    let set_total: f32 = explicit.iter().flatten().sum();
+    let set_total: f64 = explicit.iter().flatten().sum();
     let auto_cols: Vec<usize> = (0..n_cols).filter(|&c| explicit[c].is_none()).collect();
 
     if !auto_cols.is_empty() {
         // Remaining space divides EQUALLY over the auto columns (fixed layout
         // has no content information to weight by).
-        let equal = (available - set_total).max(0.0) / auto_cols.len() as f32;
+        let equal = (available - set_total).max(0.0) / auto_cols.len() as f64;
         for c in auto_cols {
             explicit[c] = Some(equal);
         }
